refactor: optimize data synchronization and standardize numeric types
- Standardized `meetingNumber` and `index` types from unsigned (`UByte`/`u8`) to signed (`Byte`/`i8`) across Rust entities, Kotlin models, and database schemas.
- Introduced granular synchronization methods `syncAttendances` and `syncAssignment` in `LmsRepository` to allow targeted updates.
- Refactored `AssignmentViewModel` and `AttendanceViewModel` to use these granular sync methods after actions, reducing unnecessary overhead from `syncAll`.
- Updated `AssignmentScreenData` and the corresponding Room DAO query to include `meetingUrl`.
- Renamed `firstLogin` to `login` in `LmsRepository` and updated `LoginViewModel` for clarity.
- Refined variable naming from `mess` to `msg` and updated parameter names in `AssignmentViewModel`.
- Reorganized and updated scraping logic in the Rust core for attendance and meeting number parsing.
- Enhanced `AssignmentDao` with `saveAll` support and specific entity update logic.

+++ app/src/main/kotlin/com/gaje48/lms/data/db/Mapper.kt
@@ -27,7 +27,7 @@ fun CourseEntity.toDomain() =
 
 fun MeetingEntity.toDomain() =
     Meeting(
-        meetingNumber = meetingNumber.toUByte(),
+        meetingNumber = meetingNumber,
         meetingUrl = meetingUrl,
     )
 
@@ -56,13 +56,13 @@ fun uniffi.lms_rust.MeetingEntity.toEntity() =
     MeetingEntity(
         url,
         courseCode,
-        number.toByte(),
+        number,
     )
 
 fun uniffi.lms_rust.AttendanceEntity.toEntity() =
     AttendanceEntity(
         courseCode,
-        index.toByte(),
+        index,
         isAttended,
     )
